Moving Crop obj to dataHandler.py
